medical-record & relations 1..*

+++ src/app/nutrient/nutrient.service.ts
@@ -1,24 +1,36 @@
-// user.service.ts
-
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
+
 import { Model } from 'mongoose';
+
 import { Nutrients } from './nutrient.entity';
 import { CreateNutrientDto } from './dto/create-ntrient.dto';
-
-
+import { Meals } from '../meals/meal.entity';
 
 @Injectable()
 export class NutrientService {
-  constructor(@InjectModel(Nutrients.name) private nutrientModel: Model<Nutrients>) {}
+  constructor(
 
-   async create(name: string, description: string, ): Promise<Nutrients> {
+    @InjectModel(Nutrients.name) private nutrientModel: Model<Nutrients>,
+    @InjectModel(Meals.name) private mealModel: Model<Meals>,
+
+  ) { }
+
+  async create(createNutrientDto: CreateNutrientDto): Promise<Nutrients> {
+    const { name, description, mealId } = createNutrientDto;
+
+    const newNutrient = new this.nutrientModel({
+      name,
+      description,
+      meal: mealId,
+    });
+
     try {
-        const newNutrient = new this.nutrientModel({ name, description });
-        return await newNutrient.save();
-
+      const nutrient = await newNutrient.save();
+      await this.mealModel.findByIdAndUpdate(mealId, { $push: { nutrients: nutrient._id } });
+      return nutrient;
     } catch (error) {
-        throw new Error('Failed to create nutrient');
+      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
   }
 
@@ -26,11 +38,11 @@ export class NutrientService {
     try {
       const nutrient = await this.nutrientModel.findOne({ name }).exec();
       if (!nutrient) {
-        throw new NotFoundException('Nutrient not found');
+        throw new HttpException("Nutrient not found", HttpStatus.BAD_REQUEST);
       }
       return nutrient;
     } catch (error) {
-      throw new Error('Failed to find nutrient by name');
+      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
   }
 
@@ -38,22 +50,23 @@ export class NutrientService {
     try {
       const nutrient = await this.nutrientModel.findById(id).exec();
       if (!nutrient) {
-        throw new NotFoundException('Nutrient not found');
+        throw new HttpException("Nutrient not found", HttpStatus.BAD_REQUEST);
       }
       return nutrient;
     } catch (error) {
-      throw new NotFoundException('Failed to get nutrient by id');
+      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
   }
 
- 
-
-  async findAll(): Promise<{ message: string, nutrients: Nutrients[] }> {
+  async findAll(): Promise<{ message: string; nutrients: Nutrients[] }> {
     try {
-      const nutrients = await this.nutrientModel.find().exec();
+      const nutrients = await this.nutrientModel.find({})
+        .populate('meal')
+        .exec();
       return { message: 'Nutrients retrieved successfully', nutrients };
     } catch (error) {
-      throw new NotFoundException('Failed to find nutrients');
+      console.error('Error retrieving nutrients:', error.message);
+      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
   }
 
@@ -61,24 +74,31 @@ export class NutrientService {
     try {
       const updatedNutrient = await this.nutrientModel.findByIdAndUpdate(id, updatenutrientDto, { new: true }).exec();
       if (!updatedNutrient) {
-        throw new NotFoundException('Nutrient not found');
+        throw new HttpException("Nutrient not found", HttpStatus.BAD_REQUEST);
       }
       return updatedNutrient;
     } catch (error) {
-      throw new Error('Failed to update nutrient');
+      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
   }
 
   async delete(id: string): Promise<{ message: string }> {
     try {
+      const nutrient = await this.nutrientModel.findById(id).exec();
+      if (!nutrient) {
+        throw new HttpException("Nutrient not found", HttpStatus.BAD_REQUEST);
+      }
+      await this.mealModel.findByIdAndUpdate(nutrient.meal, {
+        $pull: { nutrients: nutrient._id },
+      }).exec();
       const deleted = await this.nutrientModel.findByIdAndDelete(id).exec();
       if (!deleted) {
-        throw new NotFoundException('Nutrient not found');
+        throw new HttpException("Nutrient not found", HttpStatus.BAD_REQUEST);
       }
       return { message: 'Nutrient deleted successfully' };
     } catch (error) {
-      throw new Error('Failed to delete nutrient');
+      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
-  } 
+  }
 
 }
